+++ data/scripts/complexCraft/mainCaliber.lua
@@ -125,3 +125,9 @@ end
 function initialize(...)
     if mainCaliber.initialize then return mainCaliber.initialize(...) end
 end
+
+
+-- Global Event Callbacks
+function checkWeapons(...)
+    if mainCaliber.checkWeapons then return mainCaliber.checkWeapons(...) end
+end